move fallback cases to a single file
